change dtypes for new numpy

+++ hawp/ssl/datasets/synthetic_dataset.py
@@ -488,7 +488,7 @@ class SyntheticShapes(Dataset):
     @staticmethod
     def junc_to_junc_map(junctions, image_size):
         """ Convert junction points to junction maps. """
-        junctions = np.round(junctions).astype(np.int)
+        junctions = np.round(junctions).astype(np.int32)
         # Clip the boundary by image size
         junctions[:, 0] = np.clip(junctions[:, 0], 0., image_size[0]-1)
         junctions[:, 1] = np.clip(junctions[:, 1], 0., image_size[1]-1)
@@ -497,7 +497,7 @@ class SyntheticShapes(Dataset):
         junc_map = np.zeros([image_size[0], image_size[1]])
         junc_map[junctions[:, 0], junctions[:, 1]] = 1
 
-        return junc_map[..., None].astype(np.int)
+        return junc_map[..., None].astype(np.int32)
 
     def train_preprocessing(self, data, disable_homoaug=False):
         """ Training preprocessing. """
