+++ hawp/ssl/datasets/wireframe_dataset.py
@@ -675,7 +675,7 @@ class WireframeDataset(Dataset):
     @staticmethod
     def junc_to_junc_map(junctions, image_size):
         """ Convert junction points to junction maps. """
-        junctions = np.round(junctions).astype(np.int)
+        junctions = np.round(junctions).astype(np.int32)
         # Clip the boundary by image size
         junctions[:, 0] = np.clip(junctions[:, 0], 0., image_size[0]-1)
         junctions[:, 1] = np.clip(junctions[:, 1], 0., image_size[1]-1)
@@ -684,7 +684,7 @@ class WireframeDataset(Dataset):
         junc_map = np.zeros([image_size[0], image_size[1]])
         junc_map[junctions[:, 0], junctions[:, 1]] = 1
 
-        return junc_map[..., None].astype(np.int)
+        return junc_map[..., None].astype(np.int32)
     
     def parse_transforms(self, names, all_transforms):
         """ Parse the transform. """
